+++ src/CMakeLists.txt
@@ -1,1 +1,2 @@
-add_executable(TKOM main.cpp Launcher.cpp Scanner.cpp Interfaces.cpp Token.cpp Parser.cpp Visitor.cpp RepresentationConverter.cpp)
+add_executable(TKOM main.cpp Launcher.cpp Scanner.cpp Interfaces.cpp Token.cpp Parser.cpp Visitor.cpp
+        RepresentationConverter.cpp EvaluationVisitor.cpp)
